requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.0
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
Pillow>=10.0
openai>=1.0
